requirements.txt
appdirs
astroid
azure-common
azure-core
azure-identity
azure-mgmt-core
azure-mgmt-network
azure-mgmt-resource
CacheControl
certifi
cffi
chardet
colorama
contextlib2
cryptography
distlib
distro
html5lib
idna
ipaddr
isodate
isort
lazy-object-proxy
lockfile
mccabe
msal
msal-extensions
msgpack
msrest
oauthlib
packaging
pep517
portalocker
prettytable
progress
pycparser
PyJWT
pylint
pyparsing
pytoml
requests
requests-oauthlib
retrying
six
toml
urllib3
wcwidth
webencodings
wrapt
